feat: make config comments conditional, refactor some stuff
