+++ packages/config/src/devices/DeviceConfig.unit._test.ts
