+++ packages/config/src/devices/shared.ts
@@ -1,0 +1,11 @@
+export interface FirmwareVersionRange {
+	min: string;
+	max: string;
+}
+
+export interface DeviceID {
+	manufacturerId: number;
+	productType: number;
+	productId: number;
+	firmwareVersion?: string;
+}
