+++ packages/config/src/index.ts
@@ -1,6 +1,11 @@
 export * from "./ConfigManager";
 export * from "./DeviceClasses";
-export * from "./Devices";
+export * from "./devices/AssociationConfig";
+export * from "./devices/DeviceConfig";
+export * from "./devices/DeviceMetadata";
+export * from "./devices/EndpointConfig";
+export * from "./devices/ParamInformation";
+export * from "./devices/shared";
 export * from "./Indicators";
 export * from "./Logger_safe";
 export * from "./Manufacturers";
