+++ packages/config/src/utils.ts
@@ -2,7 +2,7 @@ import { formatId, padVersion } from "@zwave-js/shared";
 import * as fs from "fs-extra";
 import path from "path";
 import * as semver from "semver";
-import type { DeviceConfigIndexEntry } from "./Devices";
+import type { DeviceConfigIndexEntry } from "./devices/DeviceConfig";
 import type { ConfigLogger } from "./Logger";
 
 /** The absolute path of the embedded configuration directory */
